+++ main.tf
@@ -3,7 +3,7 @@ terraform {
   required_providers {
     aws = {
       source  = "hashicorp/aws"
-      version = "3.11.0"
+      version = "~> 3.0"
     }
   }
 }
@@ -11,6 +11,12 @@ terraform {
 provider "aws" {
   region = var.region
 }
+
+data "aws_caller_identity" "current" {}
+
+################################################################
+## S3 BUCKET
+################################################################
 
 module "s3_bucket" {
   source = "terraform-aws-modules/s3-bucket/aws"
@@ -37,14 +43,18 @@ module "s3_bucket" {
   }
 }
 
+################################################################
+## DYNAMODB
+################################################################
+
 module "dynamodb_table" {
   source = "terraform-aws-modules/dynamodb-table/aws"
 
   name           = var.dynamodb_table_name
   hash_key       = "name"
   write_capacity = var.dynamodb_write_capacity
-  read_capacity = var.dynamodb_read_capacity
-  billing_mode = var.dynamodb_billing_mode
+  read_capacity  = var.dynamodb_read_capacity
+  billing_mode   = var.dynamodb_billing_mode
 
   attributes = [
     {
@@ -54,4 +64,84 @@ module "dynamodb_table" {
   ]
 
   tags = var.dynamodb_tags
+}
+
+################################################################
+## LAMBDA
+################################################################
+
+module "lambda_function" {
+  source = "terraform-aws-modules/lambda/aws"
+
+  count = var.deploy_lambda ? 1 : 0
+
+  function_name            = var.lambda_function_name
+  description              = var.lambda_description
+  handler                  = var.lambda_handler
+  runtime                  = var.lambda_runtime
+  timeout                  = var.lambda_timeout
+  publish                  = true
+  create_role              = true
+  role_name                = var.iam_role_name
+  attach_policy_statements = true
+  policy_statements = {
+    get_dynamodb = {
+      effect    = "Allow",
+      actions   = [
+        "dynamodb:Scan"
+      ],
+      resources = ["arn:aws:dynamodb:eu-west-2:${data.aws_caller_identity.current.account_id}:table/${var.dynamodb_table_name}"]
+    },
+    s3_read = {
+      effect = "Allow",
+      actions = [
+        "s3:HeadObject",
+        "s3:GetObject",
+        "s3:ListBucket"
+      ],
+      resources = ["arn:aws:s3:::${var.bucket_name}/*"]
+    }
+  }
+
+  source_path = "lambda_code/parse_dynamodb.py"
+
+  store_on_s3 = true
+  s3_bucket   = var.bucket_name
+
+  environment_variables = {
+    region                   = var.region
+    s3_bucket                = module.s3_bucket.this_s3_bucket_id
+    table_name               = module.dynamodb_table.this_dynamodb_table_id
+    years_back               = var.lambda_years_back
+    slack_webhook            = var.slack_webhook
+    slack_channel            = var.slack_channel
+    presigned_url_expiration = var.presigned_url_expiration
+  }
+
+  tags = var.lambda_tags
+
+  depends_on = [
+    module.s3_bucket
+  ]
+}
+
+################################################################
+## CLOUDWATCH EVENTS
+################################################################
+
+resource "aws_cloudwatch_event_rule" "this" {
+  count = var.deploy_lambda ? 1 : 0
+
+  name        = var.cloudwatch_event_rule_name
+  description = var.cloudwatch_event_rule_description
+
+  schedule_expression = var.cloudwatch_event_rule_expression
+}
+
+resource "aws_cloudwatch_event_target" "this" {
+  count = var.deploy_lambda ? 1 : 0
+
+  rule      = aws_cloudwatch_event_rule.this[0].name
+  target_id = "SendToLambda"
+  arn       = module.lambda_function[0].this_lambda_function_arn
 }